backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.6.1
langchain==0.0.352
faiss-cpu==1.7.4
pdfplumber==0.10.3
python-docx==1.1.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
tiktoken==0.5.2
aiofiles==23.2.1 